1) Added parameters protein and dna responsible for switching on/off in md.cu file. To enable/disable DNA, use: dna <on/off> For protein, use: protein <on/off>
2) Added function getIndexInTOP() to topio.cpp and topio.h
Function takes input a number of atom (from TOP file) and output a serial number of the atom.
It allows not to think about from which number atom indices start in the TOP file (i.e. from 0 or from 1)

This branch currently works only with protein systems without commenting DNA in md.cu